Refactor setupCase
